backend/pyproject.toml
[project]
name = "gemini-balance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "bcrypt<4.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis[hiredis]>=6.2.0",
    "uvicorn>=0.35.0",
    "aiosqlite>=0.21.0",
]

# [[tool.uv.index]]
# url = "https://pypi.tuna.tsinghua.edu.cn/simple"
# url = "https://mirrors.aliyun.com/pypi/simple"
# default = true

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]
